requirements.txt
decorator==4.0.9
Django==1.10.3
django-registration==2.2
django-widget-tweaks==1.4.1
djangorestframework==3.5.3
psycopg2==2.6.2
virtualenv==14.0.3